requirements.txt
# Data manipulation and analysis
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2

# Data visualization
matplotlib>=3.4.2
seaborn>=0.11.1
shap>=0.39.0

# Deep learning
tensorflow>=2.6.0
keras>=2.6.0

# Time series analysis
statsmodels>=0.12.2
pmdarima>=1.8.2

# Machine learning
xgboost>=1.4.2

# Data acquisition
yfinance>=0.1.63

# Utilities
pillow>=8.2.0
tqdm>=4.61.2